+++ src/http/utils.ts
@@ -111,10 +111,13 @@ export function camelize(string: string): string {
 }
 
 export function decamelize(string: string): string {
-  return string
-    .split(/(?=[A-Z])/)
-    .join('_')
-    .toLowerCase();
+  // If all characters are uppercase, no need to decamelize as this is very likely a language code
+  return string.toUpperCase() === string
+    ? string
+    : string
+        .split(/(?=[A-Z])/)
+        .join('_')
+        .toLowerCase();
 }
 
 export function camelizeKeys(