+++ src/services/index.ts
@@ -8,5 +8,6 @@ export * from './templates';
 export * from './mails';
 export * from './dispatchers';
 export * from './payments';
+export * from './localizations';
 export * from './profiles';
 export * from './notifications';